+++ grunt/concat.js
@@ -10,8 +10,11 @@ module.exports = function (grunt) {
                 "js/libs/ui/Components/Icon/FluentIcons.css",
                 "js/libs/ui/Components/Header/Header.css",
                 "js/libs/ui/Components/Button/Button.css",
+                "js/libs/ui/Components/Button/ExpandableButton.css",
                 "js/libs/ui/Components/Spinner/Spinner.css",
-                "js/libs/ui/Components/TextField/TextField.css"
+                "js/libs/ui/Components/TextField/TextField.css",
+                "js/libs/ui/Components/Menu/Menu.css",
+                "js/libs/ui/Components/Menu/MenuButton.css"
             ],
             dest: "css/azure-devops-ui.css"
         }